treehouse now in the about view
