+++ app/styles/partials/_animations.scss
