+++ app/styles/partials/_media.scss
@@ -1,13 +1,28 @@
+.profile {
+    border-bottom: 1px solid #f3f3f3;
+    padding-bottom: 0.675rem;
+}
+
 .profile-pic {
     max-width: 200px;
+}
+
+.profile-phrase {
+    margin-top: 2.26rem;
+    line-height: 1.6rem; 
 }
 
 .language-color {
     margin-right: .28rem;
 }
 
+.tech {
+    margin-top: 1.5rem;
+    margin-bottom: 1.7rem;
+}
+
 .tech-icon {
-    max-height: 75px;
+    max-height: 63px;
     margin-bottom: 1.7rem;
 }
